+++ src/main/java/racingcar/controller/exception/DomainExceptionAdvice.java
@@ -1,5 +1,7 @@
-package racingcar.controller;
+package racingcar.controller.exception;
 
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -10,8 +12,9 @@ import racingcar.exception.CarNameLengthException;
 import racingcar.exception.NoCarsExistException;
 import racingcar.exception.PositionInvalidException;
 
+@Order(Ordered.HIGHEST_PRECEDENCE)
 @RestControllerAdvice
-public class ExceptionController {
+public class DomainExceptionAdvice {
 
     @ExceptionHandler({
             CarNameBlankException.class,
@@ -19,7 +22,7 @@ public class ExceptionController {
             NoCarsExistException.class,
             PositionInvalidException.class
     })
-    public ResponseEntity<ExceptionResponse> handle(RuntimeException e) {
+    public ResponseEntity<ExceptionResponse> handledException(final RuntimeException e) {
         return new ResponseEntity<>(
                 new ExceptionResponse(e.getMessage()),
                 HttpStatus.BAD_REQUEST
